Directory update get-users-controller and renaming get-users-controller-interface
