+++ src/controllers/get-users-controller/get-users-controller.ts
@@ -1,13 +1,13 @@
 import type { Response, Request } from 'express'
-import type { IGetUsersController } from '../../interfaces/create-user-get-controller-interface'
-import type { IUserRepository } from '../../interfaces/user-repository-interface'
+import type { IGetUsersUseCase } from '../../interfaces/get-users-use-case-interface'
+import type { IGetUsersController } from '../../interfaces/get-users-controller-interface'
 
 export class GetUsersController implements IGetUsersController {
-  constructor(private readonly repository: IUserRepository) {}
+  constructor(private readonly getUsersUseCase: IGetUsersUseCase) {}
 
   async handle(request: Request, response: Response): Promise<Response> {
     try {
-      const users = await this.repository.findAll()
+      const users = await this.getUsersUseCase.execute()
       return response.status(200).json(users)
     } catch (error) {
       return response.status(500).json({ error: 'failed to retrive users' })
